--- DotNet/Submission/Application/Models/PatientSubmissionModel.cs
@@ -1,20 +1,0 @@
-﻿using Hl7.Fhir.Model;
-using LantanaGroup.Link.Shared.Application.Converters;
-using LantanaGroup.Link.Shared.Domain.Entities;
-using System.Text.Json.Serialization;
-
-namespace LantanaGroup.Link.Submission.Application.Models
-{
-    public class PatientSubmissionModel : BaseEntityExtended
-    {
-        public string FacilityId { get; set; } = string.Empty;
-        public string PatientId { get; set; } = string.Empty;
-        public string ReportScheduleId { get; set; }
-        public DateTime StartDate { get; set; }
-        public DateTime EndDate { get; set; }
-        [JsonConverter(typeof(FhirResourceConverter<Bundle>))]
-        public Bundle PatientResources { get; set; }
-        [JsonConverter(typeof(FhirResourceConverter<Bundle>))]
-        public Bundle OtherResources { get; set; }
-    }
-}
